pyproject.toml
[project]
name = "moodmate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Muhammad Usman GM", email = "muhammadusman75425965etc@gamil.com" }
]
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]

[project.scripts]
moodmate = "moodmate:main"
mate = "moodmate.mate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
